Fixed an issue with previous flag byte size reduction

+++ tools/flag.hpp
@@ -5,7 +5,7 @@
 
 struct Flag
 {
-    uint8_t byte;
+    uint16_t byte;
     uint8_t bit;
 
     Flag() : byte(0), bit(0)
